import-zipf generates an exponential distribution based on deltas
this way, bitmap ids and profile ids are automatically sorted which should have
a very positive performance impact. The actual distribution of the underlying
data is unknown, since the exponential distribution is the distance between set
bits. It is probably possible to mathematically characterize the generated
distribution, but we have not done this yet. Right now this is just a faster way
to generate random data which probably not a uniform distribution.