use fancy swap-or-not permutation generator
Found a cool paper on a way to effectively simulate just the
swaps we care about for a shuffle algorithm to determine
what card ends up in a given position. Added testing to make
sure this actually works.

This gives constant-time lookups for the Nth item in a permutation,
and also allows iterating without storing the previous value. The
permutation is controlled by the seed, so different seeds produce
significantly different patterns. (The underlying code is using
AES hashes of the inputs to produce the outputs.)

There is some selection bias in the PRNG, which is subject to
the usual problem that `u%max` is not uniformly distributed over
the range, but for n smaller than 2^60 or so, it's small and
we probably don't care.