cleanup and performance improvements
Drop the hits/values counters, and the interface to them, because
they no longer make sense -- this algorithm never discards values
anyway.

Treat the 128-bit result of AES as a 128-bit buffer of bits, and
only regenerate it when we switch which buffer we need to look at.

In cases where the resulting bit is 0, this results in losing
the previous value of the buffer, and then possibly needing to
regenerate it next time (about 50% of the time), but trying to
avoid that turns out to make it slower, not faster, because of
additional complications and branches.