add fancy seekable PRNG-related functionality
This commit introduces a seekable PRNG implementation,
under the name `apophenia`, for use in benchmarking and tools.
The idea is to provide generators for things like arbitrary
permutations, Zipf values, and weighted bits (bits with
a specified non-50% probability of being set) which are all
seekable, allowing generating the Nth result in a series
in constant time. This gives us tools usable for creating
data sets which contain the same data regardless of the
order in which the data are created; for instance, getting
the same bits whether creating data in row-major or
column-major order.