imagine tool rough draft
The `imagine` tool offers a way to create sample databases
with defined characteristics programatically, avoiding the need
to have large actual data sets. The intent is to support a
variety of possible use cases for use in demos and benchmarks.
The tool can also serve as something of a benchmark for ingest
speed in general.

This is a fairly rough draft of the design, but has an actual
spec of some kind so you can predict what the data set should
look like. Not pictured: Significant/reliable testing.