refactor slightly: Move documentation around, mostly
Clarify and fix various documentation and comments related
to how the seekable PRNG is used, document it in the library
along with a rationale for why it exists, and so on.

I was originally going to move the OffsetFor functionality
out of apophenia entirely, but on reflection, that works
poorly for the Permute or Zipf cases, where being able to
hint that they should use a different set of values than
you'd get otherwise seems desireable.