allow spec field overrides
This adds a fancy new feature where you can have multiple
specs for a given field. The last specs specified are tried
first, with each spec providing a chance in [0,1] that it will
be used.

For instance:

```
{ name = "numbers", type = "set", max = 100, density = 0.01, },
{ name = "numbers", type = "set", density = 1.0, },
```

This results in 5% of columns set with everything from 0..100 set,
and 95% having 1% of everything from 0..100 set.