partially add timestamps, add column offset
Now available: column offset for tasks, with the special
case "append" which follows on after the highest column value
used during this run. Note, this is arguably buggy; in low-
density fields, it's the highest column value actually populated,
not the highest column value considered.

Provided partial hooks for, but not yet using, timestamps for
mutexes and sets.