add timestamps (partially implemented)
This handles the "increasing" timestamp case. It doesn't have
any kind of column-awareness or handle random timestamps, and it
doesn't generate the right timestamps -- it figures out the
timestamp increment it'd need if every bit in rows*columns were
generated, then uses that for each 1 bit generated. But for
some cases, this will likely produce reasonable values.