Pattern generation
This adds a feature to allow creation of patterned data with bitmasks
corresponding to patterns being inserted in int-type fields.

The only currently supported pattern is "triangle", which does
digits in a pattern like
	001012012301234[...]
up to a specified limit. It can also generate masks for columns
where the generated digit is equal to, equal to or greater than,
or lower than a designated value, or the first time that values
up to N were generated in each run.

For N=10, this means that equal(0) is 10/55 of the input space,
while equal(9) is 1/55 of the input space, but distinct queries
against the values masked by them will produce exactly one value
either way; on the other hand, once(9) is 10/55 of the input space,
and denotes 10 distinct values.